+++ src/include/vfs-drivers.h
@@ -34,4 +34,6 @@ vfs_driver_t *VfsPlatformFsInit(const wapp_t *wapp, const char *options,
                                 bool readonly);
 vfs_driver_t *VfsSocketInit(const wapp_t *wapp, const char *options);
 vfs_driver_t *VfsVirtualInit(const wapp_t *wapp, const char *options);
+vfs_driver_t *VfsGpioInit(const wapp_t *wapp, const char *options);
+vfs_driver_t *VfsWifiInit(const wapp_t *wapp, const char *options);
 vfs_driver_t *VfsWantedInit(const wapp_t *wapp, const char *options);